requirements.txt
streamlit
google-generativeai
fpdf
reportlab